+++ src/pages/index.js
@@ -1,4 +1,4 @@
 export { default as default } from './default.js';
 export { default as notFound } from './not-found.js';
-export { default as report } from './report.js';
+export { default as discover } from './discover.js';
 export { default as viewsShowcase } from './views-showcase.js';