+++ src/views/editor/editors.js
@@ -36,7 +36,7 @@ class Editor extends Emitter {
         super();
 
         this.el = document.createElement('div');
-        this.el.className = 'discovery-editor';
+        this.el.className = 'discovery-view-editor';
 
         const self = this;
         const cm = CodeMirror(this.el, {
@@ -140,8 +140,8 @@ class QueryEditor extends Editor {
             };
         } });
 
-        this.inputPanelEl = createElement('div', 'discovery-editor__input-panel');
-        this.outputPanelEl = createElement('div', 'discovery-editor__output-panel');
+        this.inputPanelEl = createElement('div', 'discovery-view-editor__input-panel');
+        this.outputPanelEl = createElement('div', 'discovery-view-editor__output-panel');
         this.el.append(this.inputPanelEl, this.outputPanelEl);
     }
     setValue(value, data, context) {